+++ src/lib.rs
@@ -2,3 +2,4 @@ pub mod classes;
 mod extra;
 pub mod linux;
 pub mod traits;
+pub type PCIDevice = linux::LinuxPCIDevice;
